- fixed IdArrayType initialize so it uses dtype correctly - fixed scenario to use updated attribute names for SpillContainer
SpillContainer:
- made current_time required arg for prepare_for_model_step
- added _array_types and created array_types property
  - it returns a new dict(self._array_types)
  - this is so new array_types are added only in prepare_for_model_run()
    at beginning of run
  - or user can add a new array in the middle of the run,
    with size consistent with other arrays. This will automatically
    infer the associated ArrayType and add it to _array_types dict
  - the properties of the existing ArrayTypes can be modified using
    array_types

+++ py_gnome/gnome/array_types.py
@@ -116,7 +116,7 @@ class IdArrayType(ArrayType):
     """
     def initialize(self, num_elements):
         array = np.arange(self.initial_value,
-                          num_elements + self.initial_value)
+                          num_elements + self.initial_value, dtype=self.dtype)
         return array
